requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
ipykernel
dill
PyYAML
Flask
mlflow==2.8.0
